requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
albumentations=0.5.1=pypi_0
ca-certificates=2021.7.5=h06a4308_1
cached-property=1.5.2=pypi_0
certifi=2021.5.30=py37h06a4308_0
cudatoolkit=10.0.130=0
cycler=0.10.0=py_2
cython=0.29.21=pypi_0
decorator=4.4.2=pypi_0
freetype=2.10.2=he06d7ca_0
h5py=3.1.0=pypi_0
icu=67.1=he1b5a44_0
imageio=2.8.0=pypi_0
imgaug=0.4.0=pypi_0
intel-openmp=2020.1=217
joblib=0.15.1=pypi_0
jpeg=9d=h516909a_0
kiwisolver=1.2.0=py37h99015e2_0
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libedit=3.1.20191231=h7b6447c_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.5.0=hdf63c60_6
liblapack=3.8.0=14_openblas
libopenblas=0.3.7=h5ec1e0e_6
libpng=1.6.37=hed695b0_1
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=hc7e4089_6
libwebp-base=1.1.0=h516909a_3
lz4-c=1.9.2=he1b5a44_1
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py37h1d35a4c_1
mkl=2020.1=217
ncurses=6.2=he6710b0_1
networkx=2.4=pypi_0
nibabel=3.1.0=pypi_0
ninja=1.10.0=hc9558a2_0
numpy=1.18.5=py37h8960a57_0
olefile=0.46=py_0
opencv-contrib-python=4.4.0.46=pypi_0
opencv-python=4.4.0.46=pypi_0
opencv-python-headless=4.4.0.46=pypi_0
openssl=1.1.1k=h27cfd23_0
packaging=20.4=pypi_0
pandas=1.1.1=pypi_0
pillow=7.1.2=py37h718be6c_0
pip=20.1.1=py37_1
protobuf=3.14.0=pypi_0
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.7.7=hcff3b4d_5
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.4.0=py3.7_cuda10.0.130_cudnn7.6.3_0
pytz=2020.1=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.3.1=pypi_0
readline=8.0=h7b6447c_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.23.1=pypi_0
scipy=1.5.0=pypi_0
setproctitle=1.1.10=pypi_0
setuptools=47.3.1=py37_0
shapely=1.7.1=pypi_0
simpleitk=2.0.2=pypi_0
six=1.15.0=pyh9f0ad1d_0
sklearn=0.0=pypi_0
spams=2.6.2.5=pypi_0
sqlite=3.32.3=h62c20be_0
staintools=2.1.2=pypi_0
tensorboardx=2.1=pypi_0
thop=0.0.31-2005241907=pypi_0
threadpoolctl=2.1.0=pypi_0
tifffile=2020.7.4=pypi_0
tk=8.6.10=hbc83047_0
torchstat=0.0.7=pypi_0
torchvision=0.5.0=py37_cu100
tornado=6.0.4=py37h8f50634_1
tqdm=4.54.0=pypi_0
wheel=0.34.2=py37_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h6597ccf_3
